+++ src/commands/rcon/general/teleport.ts
@@ -1,0 +1,103 @@
+import BaseRCONCommand from "../../../structures/BaseRCONCommands";
+import RCONCommandContext from "../../../structures/RCONCommandContext";
+import TeleportConfig, {
+    Coordinates,
+    Location,
+} from "../../../structures/TeleportConfig";
+import Watchdog from "../../../structures/Watchdog";
+
+export default class Teleport extends BaseRCONCommand {
+    constructor(bot: Watchdog, commandName: string) {
+        super(bot, {
+            name: commandName,
+            usage: "teleport [player name/id] ([x] [y] [z])/[location]",
+            aliases: ["tp"],
+        });
+    }
+
+    async execute(ctx: RCONCommandContext) {
+        if (!ctx.args.length) {
+            const locations: {
+                [name: string]: {
+                    aliases: string[];
+                    coordinates: Coordinates;
+                };
+            } = TeleportConfig.get(`maps.${ctx.rcon.currentMap}.locations`, {});
+
+            const locationsMessage =
+                "Locations: " +
+                Object.keys(locations)
+                    .map((name) => name)
+                    .join(", ");
+
+            return await ctx.say(
+                locationsMessage.length
+                    ? locationsMessage
+                    : "No teleport locations has been added"
+            );
+        }
+
+        let foundWithName = false;
+        const location = Object.entries<Location>(
+            TeleportConfig.get(`maps.${ctx.rcon.currentMap}.locations`, {})
+        ).find(([name, location]) => {
+            if (name === ctx.args.join(" ")) {
+                foundWithName = true;
+                return true;
+            } else if (location.aliases.includes(ctx.args.join(" ")))
+                return true;
+        });
+        if (location) {
+            const name = ctx.args
+                .slice(
+                    0,
+                    ctx.args.length -
+                        (foundWithName ? location[0] : ctx.args.join(" "))
+                            .length
+                )
+                .join(" ");
+            const player = await ctx.rcon.getIngamePlayer(name);
+            if (!player) return await ctx.say("Player not found");
+
+            return await ctx.rcon.teleportPlayer(
+                ctx.player.id,
+                location[1].coordinates
+            );
+        }
+
+        let player: { id: any; name?: string };
+        if (ctx.args.length > 3) {
+            const name = ctx.args
+                .slice(
+                    0,
+                    ctx.args.length -
+                        (foundWithName ? location[0] : ctx.args.join(" "))
+                            .length
+                )
+                .join(" ");
+            player = await ctx.rcon.getIngamePlayer(name);
+            if (!player) return await ctx.say("Player not found");
+            console.log(player);
+        }
+
+        // X coordinate
+        let x: string | number = ctx.args[ctx.args.length - 3];
+        if (!x) return;
+        x = parseInt(x);
+        if (isNaN(x)) return;
+
+        // Y coordinate
+        let y: string | number = ctx.args[ctx.args.length - 2];
+        if (!y) return;
+        y = parseInt(y);
+        if (isNaN(y)) return;
+
+        // Z coordinate
+        let z: string | number = ctx.args[ctx.args.length - 1];
+        if (!z) return;
+        z = parseInt(z);
+        if (isNaN(z)) return;
+
+        await ctx.rcon.teleportPlayer(player?.id || ctx.player.id, { x, y, z });
+    }
+}