+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.0] - 2021-08-29
+
+### Added
+
+-   Added a teleport system (no teleport locations are specified by default)
+-   Also added 3 new discord slash commands: `teleportadd`, `teleportremove` and `teleportedit`
+
+### Changed
+
+-   Made ingame commands array could be empty (essentially disabling ingame commands)
+-   Moved save admin list to run when keepalive succeeds instead of also running after keepalive error
+
 ## [1.12.1] - 2021-08-28
 
 ### Fixed